Swaps out development fetchUrl for the real deal
