+++ src/cmd/url.js
@@ -2,10 +2,9 @@
 
 const log = require('../utils/log.js');
 const config = require('../utils/config.js');
-const { prompt } = require('../utils/prompt.js');
-
 const finish = require('../utils/finish.js');
-const filter = require('../utils/filter')
+const filter = require('../utils/filter');
+const { taskPrompt } = require('../utils/prompt')
 const opn = require('opn');
 
 
@@ -18,7 +17,7 @@ let OPEN = false;
  * @param args index
  * @param env
  */
-function action(args, env) {
+async function action(args, env) {
 
   // Reset URLs
   URLS = [];
@@ -26,124 +25,68 @@ function action(args, env) {
   // Set Open flag
   OPEN = env.open === undefined ? false : env.open;
 
+  const user = config.user(user => user)
+
+  let indices = []
+
   // Prompt for task
   if ( args.length < 1 ) {
-    prompt('Task:', _promptFinished);
+    indices = taskPrompt('Task:')
+    args[0] = indices
   }
 
+
+  log.spinner.start("Getting Task(s)");
   // Use provided args
-  else {
-    for ( let i = 0; i < args[0].length; i++ ) {
-      _process(args[0][i], i+1, args[0].length);
-    }
+  for (const arg in args[0]) {
+      if (Object.hasOwnProperty.call(args[0], arg)) {
+          const index = args[0][arg];
+          const filterString = filter('hasUrl:true');
+          let task =  await user.tasks.rtmIndexFetchTask(index,filterString)
+          if (task.err == undefined ) {
+              task = task.task
+          } else {
+            log.spinner.warn('Task #' + index + ' does not have a URL or is not found');
+          }
+
+          // Push to URLS
+          if ( task && task.url !== undefined ) {
+            URLS.push({
+            index: index,
+            url: task.url,
+            name: task.name
+            });
+          }         
+      }
   }
-
-}
-
-
-/**
- * Process the returned answers
- * @private
- */
-function _promptFinished(answers) {
-  for ( let i = 0; i < answers.length; i++ ) {
-    let answer = answers[i];
-    _process(answer[0], i+1, answers.length);
-  }
-}
-
-
-/**
- * Process the request
- * @private
- */
-function _process(index, count=1, max=1) {
-
-  // Display info
-  log.spinner.start("Getting Task(s)...");
-
-  // Get User
-  config.user(function(user) {
-
-    // Parse arguments
-    index = parseInt(index.trim());
-
-    // Get Task
-    const filterString = filter("hasUrl:true");
-    user.tasks.getTask(index, filterString, function(err, task) {
-      if ( err ) {
-        if ( err.code === -3 ) {
-          log.spinner.warn("Task #" + index + " does not have a URL or is not found");
-        }
-        else {
-          log.spinner.error("Could not get Task #" + index + " (" + err.msg + ")");
-        }
+  // Print URLs
+  log.spinner.stop();
+  for ( let i = 0; i < URLS.length; i++ ) {
+      log(URLS[i].index + " " + URLS[i].name + " " + URLS[i].url);
       }
 
-      // Add URL
-      else if ( task.url !== undefined ) {
-        URLS.push({
-          index: index,
-          url: task.url,
-          name: task.name
-        });
-      }
-
-      // Finish
-      _processFinished(count, max);
-
-    });
-
-  });
-
-}
-
-/**
- * Request Callback
- * @private
- */
-function _processFinished(count, max) {
-  log.spinner.start("Getting Task [" + count + "/" + max + "]...");
-
-  // All tasks returned...
-  if ( count === max ) {
-    log.spinner.stop();
-
-    // Print URLs
+  // Open URL
+  if ( OPEN ) {
     for ( let i = 0; i < URLS.length; i++ ) {
-      console.log(URLS[i].index + " " + URLS[i].name + " " + URLS[i].url);
-    }
-
-    // Open URL
-    if ( OPEN ) {
-      for ( let i = 0; i < URLS.length; i++ ) {
-        opn(URLS[i].url, {wait: false}).then(function() {
-          _finish(i, URLS.length);
-        });
-      }
-    }
-    else {
-      return finish();
+      opn(URLS[i].url, {wait: false})//.then(function() {});
     }
   }
+  
 
-  function _finish(count, max) {
-    if ( count === max-1 ) {
-      return finish();
-    }
-  }
+  finish()
+  
 
 }
 
 
 module.exports = {
-  command: 'url [indices...]',
-  options: [
-    {
-      option: "-o, --open",
-      description: "Open the URLs in a browser"
-    }
-  ],
-  description: 'Display the associated URL of a Task',
-  action: action
-};
+    command: 'fetchUrl [indices...]',
+    options: [
+      {
+        option: "-o, --open",
+        description: "Open the URLs in a browser"
+      }
+    ],
+    description: 'Display the associated URL of a Task',
+    action: action
+  };
